GFORMS-2763 - Determine how form builder could be deployed as part of gform-frontend rather than extension
